[BACK-792] Provide a more user friendly error message when creating a duplicate story
A custom error message is now returned if a story with a certain URL
 is already present in a collection. This message is returned on both
adding a new story with a duplicate URL and trying to update an existing
story and set its URL to one that matches another story in a collection.

Added quotes to error message about duplicate slugs for collections
to make it the match the format for duplicate author slugs.

Updated/added tests.

Updated README with Docker command for running integration tests.