[BACK-755] Add an extra partnership query
The frontend is having a hard time refreshing the partnership component if
the CollectionPartnerAssociation entity is queried by the value of its 'externalId'
field.

To fix this, I've added a new query getCollectionPartnerAssociationForCollection
- this one requests a collection external id rather the association's own external id.

I have left the original getCollectionPartnerAssociation intact even though it's currently
not used anywhere on the frontend.